+++ Tehtävä03/src/Laiteosa.java
@@ -3,5 +3,5 @@
  * @author stripeswamp
  */
 public interface Laiteosa {
-    float getHinta();
+    double getHinta();
     }
