+++ Tehtävä03/src/Verkkokortti.java
@@ -3,7 +3,7 @@
  * @author stripeswamp
  */
 public class Verkkokortti implements Laiteosa {
-    private float hinta;
+    private double hinta;
     
     public Verkkokortti(float inputHinta)
     {
@@ -11,7 +11,7 @@ public class Verkkokortti implements Laiteosa {
     }
     
     @Override
-    public float getHinta()
+    public double getHinta()
     {
         return this.hinta;
     }
